requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.9.0
fsspec==2023.4.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy~=1.26.3
pandas==2.1.4
Pillow~=10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
six==1.16.0
sympy==1.12
torch~=2.1.2
torchaudio==2.1.2+cpu
torchvision~=0.16.2
typing_extensions==4.4.0
tzdata==2023.4
Unidecode==1.3.7
urllib3==1.26.13

matplotlib~=3.8.2
